+++ zxing-android-embedded/src/com/google/zxing/integration/android/IntentIntegrator.java
@@ -34,7 +34,6 @@ import java.util.List;
 import java.util.Map;
 
 /**
- *
  * @author Sean Owen
  * @author Fred Lin
  * @author Isaac Potoczny-Jones
@@ -81,7 +80,7 @@ public class IntentIntegrator {
 
     private final Activity activity;
     private android.app.Fragment fragment;
-    private android.support.v4.app.Fragment supportFragment;
+    private androidx.fragment.app.Fragment supportFragment;
 
     private final Map<String, Object> moreExtras = new HashMap<>(3);
 
@@ -137,7 +136,7 @@ public class IntentIntegrator {
      *                 {@link #startActivityForResult(Intent, int)} will be called on the {@link Fragment} instead
      *                 of an {@link Activity}
      */
-    public static IntentIntegrator forSupportFragment(android.support.v4.app.Fragment fragment) {
+    public static IntentIntegrator forSupportFragment(androidx.fragment.app.Fragment fragment) {
         IntentIntegrator integrator = new IntentIntegrator(fragment.getActivity());
         integrator.supportFragment = fragment;
         return integrator;
@@ -265,6 +264,7 @@ public class IntentIntegrator {
     /**
      * Initiates a scan for all known barcode types with the default camera.
      * And starts a timer to finish on timeout
+     *
      * @return Activity.RESULT_CANCELED and true on parameter TIMEOUT.
      */
     public IntentIntegrator setTimeout(long timeout) {
@@ -344,7 +344,7 @@ public class IntentIntegrator {
     /**
      * <p>Call this from your {@link Activity}'s
      * {@link Activity#onActivityResult(int, int, Intent)} method.</p>
-     *
+     * <p>
      * This checks that the requestCode is equal to the default REQUEST_CODE.
      *
      * @param requestCode request code from {@code onActivityResult()}
@@ -364,8 +364,8 @@ public class IntentIntegrator {
     /**
      * Parse activity result, without checking the request code.
      *
-     * @param resultCode  result code from {@code onActivityResult()}
-     * @param intent      {@link Intent} from {@code onActivityResult()}
+     * @param resultCode result code from {@code onActivityResult()}
+     * @param intent     {@link Intent} from {@code onActivityResult()}
      * @return an {@link IntentResult} containing the result of the scan. If the user cancelled scanning,
      * the fields will be null.
      */
@@ -383,9 +383,10 @@ public class IntentIntegrator {
                     rawBytes,
                     orientation,
                     errorCorrectionLevel,
-                    barcodeImagePath);
+                    barcodeImagePath,
+                    intent);
         }
-        return new IntentResult();
+        return new IntentResult(intent);
     }
 
     private static List<String> list(String... values) {
@@ -409,6 +410,18 @@ public class IntentIntegrator {
                 intent.putExtra(key, (Float) value);
             } else if (value instanceof Bundle) {
                 intent.putExtra(key, (Bundle) value);
+            } else if (value instanceof int[]) {
+                intent.putExtra(key, (int[]) value);
+            } else if (value instanceof long[]) {
+                intent.putExtra(key, (long[]) value);
+            } else if (value instanceof boolean[]) {
+                intent.putExtra(key, (boolean[]) value);
+            } else if (value instanceof double[]) {
+                intent.putExtra(key, (double[]) value);
+            } else if (value instanceof float[]) {
+                intent.putExtra(key, (float[]) value);
+            } else if (value instanceof String[]) {
+                intent.putExtra(key, (String[]) value);
             } else {
                 intent.putExtra(key, value.toString());
             }